Unify container operations
