+++ test/test_container_arguments.hpp
@@ -1,6 +1,7 @@
 #pragma once
 #include <catch.hpp>
 #include <argparse.hpp>
+#include <test_utility.hpp>
 
 TEST_CASE("Parse vector of arguments", "[vector]") {
   argparse::ArgumentParser program("test");
@@ -24,8 +25,8 @@ TEST_CASE("Parse list of arguments", "[vector]") {
 
   auto inputs = program.get<std::list<std::string>>("input");
   REQUIRE(inputs.size() == 2);
-  REQUIRE(argparse::get_from_list(inputs, 0) == "rocket.mesh");
-  REQUIRE(argparse::get_from_list(inputs, 1) == "thrust_profile.csv");
+  REQUIRE(testutility::get_from_list(inputs, 0) == "rocket.mesh");
+  REQUIRE(testutility::get_from_list(inputs, 1) == "thrust_profile.csv");
 }
 
 TEST_CASE("Parse list of arguments with default values", "[vector]") {
@@ -38,11 +39,11 @@ TEST_CASE("Parse list of arguments with default values", "[vector]") {
 
   auto inputs = program.get<std::list<int>>("--input");
   REQUIRE(inputs.size() == 5);
-  REQUIRE(argparse::get_from_list(inputs, 0) == 1);
-  REQUIRE(argparse::get_from_list(inputs, 1) == 2);
-  REQUIRE(argparse::get_from_list(inputs, 2) == 3);
-  REQUIRE(argparse::get_from_list(inputs, 3) == 4);
-  REQUIRE(argparse::get_from_list(inputs, 4) == 5);
+  REQUIRE(testutility::get_from_list(inputs, 0) == 1);
+  REQUIRE(testutility::get_from_list(inputs, 1) == 2);
+  REQUIRE(testutility::get_from_list(inputs, 2) == 3);
+  REQUIRE(testutility::get_from_list(inputs, 3) == 4);
+  REQUIRE(testutility::get_from_list(inputs, 4) == 5);
   REQUIRE(program["--input"] == std::list<int>{1, 2, 3, 4, 5});
 }
